requirements.txt
Django
psycopg2-binary
djangorestframework
django-simple-history
drf-yasg
redis
celery[redis]
zaproxy
django-filter
coverage